[otbn] Copy over vectorized instructions: bn.addv{m}, bn.subv{m}, bn.shv and bn.trn from PQ working repo
This commit manually copied over the RTL changes in OTBN to support
bn.addv{m}, bn.subv{m}, bn.shv and bn.trn from the paper "Towards
ML-KEM and ML-DSA on OpenTitan" (https://eprint.iacr.org/2024/1192),
branch multi-cycle-bnmulv (https://github.com/PQC-OpenTitan/towards-ml-kem-and-ml-dsa-on-opentitan-dev-multi-cycle-bnmulv),
with clean-ups. All the DV-related tests are not copied over:
dv/uvm/vseq and dv/sv since we don't need them for now. Similar
for some ASIC-synthesis files with Genus.

Note that the RTL changes are not RELATED TO bn.mulv{m}, which means
only BN-ALU is modified and not BN-MAC.

The only two tests that test if the RTL is working correctly is:
- hw/ip/otbn/dv/smoke/run_smoke_isaext.sh that runs all the tests
in sw/otbn/crypto/tests/isa_ext, including bn.{add,sub}v{m}, bn.shv,
and bn.trn{1,2}, and verifies if the RTL behavior matches that of
the ISS. We don't compare the actual results with expected results.
By "passed" it means that the RTL tracing matches ISS. These tests PASSED!
- The chip-level test sw/device/tests/otbn_isa_ext_test tests only
bn.{add,sub}v{m}, not bn.shv and bn.trn. Can be run with the command:
./bazelisk.sh test --test_timeout=10000 --test_output=streamed //sw/device/tests:otbn_isa_ext_test_sim_verilator
This test also PASSED!

All the files noted with the comment "Modified by phamhnh" are just
for reminding that they were majorly rewritten, so that we can a proper
header for our paper later.