[otbn] Copy over multi-cycle bn.mulv{m}{l} from PQ working repo
This commit manually copied over the RTL changes in OTBN to support
"multi-cycle" bn.mulv{m}{l} from the paper "Towards ML-KEM and ML-DSA
on OpenTitan" (https://eprint.iacr.org/2024/1192), branch multi-cycle-bnmulv
(https://github.com/PQC-OpenTitan/towards-ml-kem-and-ml-dsa-on-opentitan-dev-multi-cycle-bnmulv),
with clean-ups. All the DV-related tests are not copied over
dv/uvm/vseq and dv/sv since we don't need them for now. Similar for
some ASIC-synthesis files with Genus.

Note that the RTL changes are ONLY for MULTI-CYCLE bn.mulv{m}{l}, which
reuses most of resources in BN-MAC unit.

The only two tests that test if the RTL is working correctly are:
- hw/ip/otbn/dv/smoke/run_smoke_isaext.sh, which now also runs
bnmulv, bnmulvm, bnmulvl, bnmulvml tests, and verifies if the RTL
behavior matches that of the ISS. We don't compare the actual results
with expected results. By "passed" it means that the RTL tracing matches
ISS. These tests PASSED!
- The chip-level test sw/device/tests/otbn_isa_ext_test now also tests
bnmulv, bnmulvl, bnmulvm, bnmulvml. Can be run with the command:
./bazelisk.sh test --test_timeout=10000 --test_output=streamed //sw/device/tests:otbn_isa_ext_test_sim_verilator
This test also PASSED!