[mldsa] Fix a register typo.
A one-character typo loaded into a WDR with the wrong register (t5 instead of
t2). It so happens that in the current code, t5 is coincidentally set to the
correct constant value (5) by previous functions. But if the code changes, it
doesn't work. This one had me scratching my head staring at debug traces for
quite a while!

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>