[otbn/dv,smoke] Add smoke tests for bn.mulv{l}.
The smoke test simulates the RTL in Verilator, collects the traces
and compares to those of the ISS. If there is no difference in these
traces, the behavior of RTL matches that of the ISS. The test then
compares the actual results with the expected results in _expected.s.

To run the smoke tests, run "hw/ip/otbn/dv/smoke/run_smoke_bnmulv.sh -h"
for more information.