[otbn] Implemented AppIntf for OTBN to KMAC connection
Added WSR and CSR registers for cfg, partial word, status, msg, and digest.
Added stall behavior to the OTBN controller for digest reads and partial word
writes when appropriate.
Updated the otbnsim model to include a partial write WSR for the KMAC msg WSR.
Updated the otbnsim model to reflect a new FIFO approach for the KMAC msg that
includes a byte-masking scheme.
Added the WSR and CSR registers to wsr.yml and csr.yml

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>