[hw/top-earlgrey] Add CW310 bitstreams for top_earlgrey BNMULV_VER{1,2,3}
These bitstreams are built for this commit:
3b2e0d149d605a87ce18f9b20a15d8789d3b12de

To load the btistream onto CW310 FPGA, run:
- ./bazelisk.sh run //sw/host/opentitantool -- fpga load-bitstream \
  hw/top_earlgrey/bitstream_cw310/bnmulv_ver*/lowrisc_systems_chip_earlgrey_cw310_0.1.bit

To run ML-KEM and ML-DSA chip-level tests with this bitstream, run:
- ./bazelisk.sh test --define bitstream=skip --test_output=streamed \
  --copt="-DKYBER_K={2,3,4}" --copt="-DBNMULV_VER={1,2,3}" \
  //sw/device/tests:otbn_mlkem_test_fpga_cw310_test_rom_ver{1,2,3}
- ./bazelisk.sh test --define bitstream=skip --test_output=streamed \
  --copt="-DDILITHIUM_MODE={2,3,5}" --copt="-DBNMULV_VER={1,2,3}" \
  //sw/device/tests:otbn_mldsa_test_fpga_cw310_test_rom_ver{1,2,3}

Note that by loading the bitstream onto the FPGA before running the
tests, the bnmulv_ver in the tag "_fpga_cw310_test_rom_ver*" doesn't
have to match that in "--copt". These two must only match if we pass
"--define bitstream=vivado" instead of "--define bitstream=skip" to
build the bitstream directly in the bazel test command (which means
the bitstream doesn't have to be reloaded onto the FPGA).