[otbn,dv/smoke] Update run_smoke_bnmulv and run_pqc to include csa_carry4 as adder to BN-{ALU,MAC}
Simulation with Verilator with CSA_CARRY4 still doesn't work due to
"Verilated didn't converge" error. We believe this is a Verilator
problem, not a design fault. To solve this, our (temporary) solution
is to set "--converge-limit=200" while otbn_top_sim is simulated with
csa_carry4 as adder in the scripts above. Similarly, to run the fusesoc
command manually, please add "--flag +csa_carry4".