[otbn,rtl] Add TOWARDS_KMAC macro
TOWARDS_KMAC is set to True in otbn_top_sim.core, all synth targets
of otbn.core, and synth targte in chip_earlgrey_cw310.core as we don't
have a new KMAC interface for now. It is only not defined for "sta"
target in otbn.core so we can obtain synth numbers for OTBN without
KMAC interface.