[otbn,mlkem] Remove cond sub in NTT VER2
NTT is used in the following parts:
- In keypair, NTT(skpv) --> pack_sk
- In keypair, NTT(e) --> pk + e --> pack_pk
- In encap, NTT(sp) --> basemul: v = sp * pkpv
- In decap, NTT(b) --> basemul: v = v * b
If we remove conditional subtraction in Montgomery multiplication in NTT,
the coefficient can grow to 2q. Then when it is subtracted from another
coeff in [0,q), the result can be negative, represented as (mod 2**16).
The modular multiplications with this result from then on will be wrong,
because of this extra mod 2**16 term. So in order to deal with this, we
switch the modulus when appropriate.

The input of NTT is [0,q). After the first layer, b*root isi n [0,2q).
The subtraction a - b*root is in (-2q,q) and addition a + b*root is in
[0,3q). With addvm and MOD = q, the results are (-q,q) and [0,2q) respectively.
This -q is the problem. Now, if we switch MOD = 2q, the results would be
in [0,2q) and [0,2q). Then from the second layer, when a and b*root are both
in [0,2q), their sum and subtraction would be in [0,4q) and (-2q,2q).
With addvm and MOD = 2q, they would be [0,2q) and [0,2q) respectively.
The result of NTT would be in [0,2q).

Thus, the idea is to switch MOD to 2q before NTT and switch it back to q
once done.

For encap/decap, after NTT, basemul computes a0*b0 + a1*b1*root and a0*b1 + a1*b0.
It doesn't matter if the inputs are in [0,2q) because the multiplications
are with conditional subtraction in the end, ensuring the results to be in [0,q).
For keypair, after NTT(skpv) is pack_sk, which requires inputs to be in [0,q). So
we need addvm for skpv before packing. Similarly to NTT(e). Hence, while
performance of encap/decap is improved, that of keypair is worse, but only
maximum of 221 cycles for mlkem1024, since we need to the coeffs again
and addvm them with 0 then save them back to the memory. In theory, the perf
of keypair is also improved if we put addvms inside the NTT, but these are
not necessary for basemul. Thus, to reduce perf of encap/decap and keeps
the same code size, we sacrifice perf of keypair, but very little. The most
important result from this is we have shown that cond_sub VER3 is not needed.