[otbn,mlkem] Remove cond sub in basemul VER2
In basemul, originally, we compute the following products:
- a0*b0 --> [0,q)
- a1*b1 --> [0,q)
- (a1*b1)*root --> [0,q)
- a0*b1 --> [0,q)
- a1*b0 --> [0,q)
Then we compute the following sums:
- a0*b0 + a1*b1*root --> [0,q)
- a0*b1 + a1*b0 --> [0,q)

Now if we remove final conditional subtraction after Montgomery
multiplication, we have:
- a0*b0, a1*b1*root, a1*b0, a0*b1 in [0,2q)
And thus the respective sums would be in [0,3q). After accumulation
for K times, the output of vector-vector product can grow to [0,9q),
which is too much for INTT.

So the idea is to switch MOD to 2q as in INTT/NTT and remove final
conditional subtraction. Then:
- a0*b0 + a1*b1*root --> [0,2q)
- a1*b0 + a0*b1 --> [0,2q)
- Accumulation: [0,2q) + [0,2q) --> [0,2q)
In the end, the outputs of vector-vector product would be in [0,2q).
For encap/decap, this is input to INTT with MOD = 2q, which is fine.
For keypair, this is input to poly_tomont with MOD = q, which means
the results are reduced mod q, correct.

Hence, we save 20*2 instructions per basemul.