[otbn] Fixed DV errors related to small message sizes in otbnsim and bad_load_store gen.
When the message was oversized with an initial length of < 8 Bytes there is a
subsequent partial write that is correctly dropped in hardware, but the model
remained stalled. The app intf flushing logic has been reworked inside the model
to recognized pending write requests in relation to flushing the App FIFO.
Additionally, discovered a case where 24B of pending data in the MSG FIFO to be
absorbed resulted in the wrong flushing delay to be added in the DV digest response.
A modulo operation had been incorrectly used and has been replaced.
Added a break to the bad_load_store loop to prevent updating the reg val
after finding a suitable target address and immediate.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>