[otbnsim] Move KMAC interface logic out of digest_ready().
Previously, some KMAC logic had to be executed in digest_ready() so that it
would happen even if OTBN was stalled waiting for KMAC to be done. However,
this meant it was not easy to make KMAC work in the background when OTBN was
not specifically waiting for KMAC and pinging the digest register. Now, the
logic runs on every clock cycle.

This change logically alters the behavior of the KMAC interface; now, it
actively checks on each step if the digest register might need more data and
tries to shift the data in or generate new data if so. This should
significantly accelerate ML-KEM and ML-DSA routines.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>