[otbn] Updated AppIntf to speculatively request the next digest after reads
Replaced the existing app_req.next logic on the AppIntf to always request
the next digest to be shifted when reading from a valid digest in the WSR.
This removes the need to stall between bn.wsrr instructions addressing
the KMAC_DIGEST WSR.

Created a counter for SHA mode to control speculation for SHA vs SHAKE
config modes. Updated driver to invalidate signals during a rsp delay
if the speculation was not needed and hold is dropped. This models
KMAC returning to Idle state after a manual run request.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>